Log volume process failure event to pod.
When volume manager or attach/detach controller start to process
a volume and encounter errors, there is no error event recorded
in Pod so that user have to wait until attach time out to realize
that something wrong with the volume. Besides, the time out event
does not show the detail error information to the user.

Need to log volume process failure event to pod to show user the
detail volume handling error immediately once the error happens.